Crate fontawesome

The Font Awesome icon set for Rust.

use fontawesome::Icon;

println!("Hello {}", Icon::Rust);

Structs

ParseVersionError

An error which can be returned when parsing a Version.

TryFromCharError

The error type returned when a character conversion fails.

Version

A version number for the Font Awesome icon set.

Enums

Icon

The enum of Font Awesome icons.

Plan

Subscription plans.

Constants

VERSION

Version 6.1.1.